+++ .changeset/move-next-builder.md
@@ -1,0 +1,5 @@
+---
+"@workflow/next": patch
+---
+
+Move NextBuilder to @workflow/next package
