+++ packages/next/src/runtime.ts
@@ -1,4 +1,4 @@
 // re-export runtime as stub for resolving to not
 // require @workflow/core be a dependency as well as
 // @workflow/next
-export * from '@workflow/cli/dist/lib/runtime';
+export * from '@workflow/core/dist/runtime';
